% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{available}
\alias{available}
\title{Identify ClimateNA data available via this package}
\usage{
available(type)
}
\arguments{
\item{type}{character. denotes the table in the sqlite database to access.
one of 'historical', 'future', 'historical_normals', or 'future_normals'.}
}
\description{
Identify ClimateNA data available via this package
}
\examples{
historical_years <- available("historical")[["years"]]
hist_nrm_prds <- available("historical_normals")[["periods"]]

future_gcms <- available("future")[["gcms"]]
future_ssps <- available("future")[["ssps"]]
future_years <- available("future")[["years"]]
future_nrm_prds <- available("future_normals")[["periods"]]
}
