% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimateNA_useTiles.R
\name{buildClimateMosaics}
\alias{buildClimateMosaics}
\title{Build climate mosaic rasters from ClimateNA tiles}
\usage{
buildClimateMosaics(
  tile,
  climVar,
  climYrs,
  srcdir = NULL,
  dstdir = NULL,
  cl = NULL
)
}
\arguments{
\item{tile}{integer. tile id (e.g., the output of \code{tileID()}).}

\item{climVar}{character. climate variables to use to construct mosaics.}

\item{climYrs}{integer. data years to use to assemble mosaic layers.}

\item{srcdir, dstdir}{character. vector of directory paths to source and destination raster files.}

\item{cl}{cluster object. if \code{NULL} (default), a cluster will be created using up to
\code{length(climVar)} number of CPU cores for parallel computation.
Users can pass their own \code{cl} object or specify option
\code{parallelly.availableCores.fallback} to reduce the number of cores used.
See \code{?parallelly.options}.}
}
\value{
character. filepaths corresponding to the raster mosaic files for each \code{climVar}.
invoked for the side effect of creating these raster mosaics on disk.
}
\description{
Build climate mosaic rasters from ClimateNA tiles
}
