% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVars.R
\name{calcStackLayersType}
\alias{calcStackLayersType}
\title{Determine the type of climate variables}
\usage{
calcStackLayersType(stacks, layers)
}
\arguments{
\item{stacks}{named list of \code{SpatRaster} objects.
names should correspond to \verb{<type>_<year>} or \verb{<type>_<period>},
where \code{type} is one of "historic" or "future"; and \code{year} or \code{period}
correspond to a climate year or climate period, respectively.}

\item{layers}{layer names (matching ClimateNA variable names) to check for in
each element of \code{stacks}.}
}
\value{
character string indicating the type (one of "historic" or "future")
}
\description{
Determine the type of climate variables
}
