% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_models.R
\name{check_models}
\alias{check_models}
\title{Check models
Check which models and which RCP scenarios are available for download}
\usage{
check_models()
}
\value{
data.frame
}
\description{
Check models
Check which models and which RCP scenarios are available for download
}
\examples{
library(climatedata)
check_models()
}
