% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVars.R
\name{checkCalcStackLayers}
\alias{checkCalcStackLayers}
\title{Check layer names within a list of raster stacks}
\usage{
checkCalcStackLayers(stacks, layers)
}
\arguments{
\item{stacks}{named list of \code{SpatRaster} objects.
names should correspond to \verb{<type>_<year>} or \verb{<type>_<period>},
where \code{type} is one of "historic" or "future"; and \code{year} or \code{period}
correspond to a climate year or climate period, respectively.}

\item{layers}{layer names (matching ClimateNA variable names) to check for in
each element of \code{stacks}.}
}
\description{
Check layer names within a list of raster stacks
}
