% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepClimateLayers.R
\name{climateBuilderDigest}
\alias{climateBuilderDigest}
\title{Digest of the bodies of the layer-building functions}
\usage{
climateBuilderDigest(envir = topenv())
}
\arguments{
\item{envir}{environment in which to find \code{.climateBuilders}.}
}
\value{
a digest string.
}
\description{
Digest of the bodies of the layer-building functions
}
\keyword{internal}
