% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateData-package.R
\docType{package}
\name{climateData-package}
\alias{climateData}
\alias{climateData-package}
\title{climateData: Utilities For Working With North American Climate Data}
\description{
Utilities for working with North American climate data from 'ClimateNA' \url{https://climatena.ca/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PredictiveEcology/climateData}
  \item Report bugs at \url{https://github.com/PredictiveEcology/climateData/issues}
}

}
\author{
\strong{Maintainer}: Alex M Chubaty \email{achubaty@for-cast.ca} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID})

Authors:
\itemize{
  \item Alex M Chubaty \email{achubaty@for-cast.ca} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID})
  \item Ian Eddy \email{ian.eddy@nrcan-rncan.gc.ca} (\href{https://orcid.org/0000-0001-7397-2116}{ORCID})
  \item Eliot McIntire \email{eliot.mcintire@nrcan-rncan.gc.ca} (\href{https://orcid.org/0000-0002-6914-8316}{ORCID})
}

}
\keyword{internal}
