% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimateNA_useTiles.R
\name{climateMosaicsParallel}
\alias{climateMosaicsParallel}
\alias{climateMosaicsNormalsParallel}
\title{Build climate mosaic rasters from ClimateNA tiles}
\usage{
climateMosaicsParallel(y, climVars, tile, srcdir, dstdir)

climateMosaicsNormalsParallel(p, climVars, tile, srcdir, dstdir)
}
\arguments{
\item{y}{integer specifying a single data year.}

\item{climVars}{character. CLimateNA climate variables to use.
see \url{https://climatena.ca/Help2#_var} for full list of variables.}

\item{tile}{integer. tile id (e.g., the output of \code{tileID()}).}

\item{srcdir, dstdir}{character. vector of directory paths to source and destination raster files.}

\item{p}{character string specifying the period.}
}
\description{
Internal functions run in parallel by \code{buildClimateMosaics()} and
\code{buildClimateMosaicsNormals()} to create the raster mosaics.
For best performance, it's arguments should be small to minimize serialization overhead
between the main R process and the workers.
}
\keyword{internal}
