% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimateNA.R
\name{ClimateNA_path}
\alias{ClimateNA_path}
\title{Construct target output paths for ClimateNA data}
\usage{
ClimateNA_path(
  dataPath,
  tile = NULL,
  type = NULL,
  msy = NULL,
  gcm = NULL,
  ssp = NULL
)
}
\arguments{
\item{dataPath}{character. path to top-level directory for reading/writing ClimateNA data.}

\item{tile}{integer. tile id (e.g., the output of \code{tileID()}).}

\item{type}{character. one of 'normals', 'historic', or 'future'.}

\item{msy}{character. one of 'MSY', 'M', or 'Y', corresponding to all, monthly, or yearly data.}
}
\value{
character string representing the path to the target output directory
}
\description{
Construct target output paths for ClimateNA data
}
