% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimateNA_createTiles.R
\name{ClimateNA_sql}
\alias{ClimateNA_sql}
\alias{checksums_sql}
\title{Connect to ClimateNA tile and checksums database}
\usage{
ClimateNA_sql(dbfile, type)

checksums_sql(dbfile, type)
}
\arguments{
\item{dbfile}{path to \code{.sqlite} database file}

\item{type}{character. denotes the table in the sqlite database to access.
one of 'historical', 'future', 'historical_normals', or 'future_normals'.}
}
\value{
named list containing:
- \code{db}, the database connection;
- \code{df}, a lazy \code{tbl_dbi} object (see \code{dbplyr}).
}
\description{
Connect to ClimateNA tile and checksums database
}
