% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimateNA_createTiles.R
\name{ClimateNA_tiles_sqlite}
\alias{ClimateNA_tiles_sqlite}
\alias{ClimateNA_tiles_gpkg}
\title{Paths to ClimateNA tile data}
\usage{
ClimateNA_tiles_sqlite()

ClimateNA_tiles_gpkg()
}
\value{
character. file path.
}
\description{
\itemize{
\item \code{ClimateNA_tiles_sqlite()}: path to a SQLite database used to look up e.g. urls to climate data
for each tile.
\item \code{ClimateNA_tiles_gpkg()}: path to a GeoPackage file defining the tile locations and extents.
}
}
