% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimateNA_useTiles.R
\name{climateStacksByYear}
\alias{climateStacksByYear}
\alias{climateStacksByPeriod}
\title{Climate Stacks by Year}
\usage{
climateStacksByYear(tifs, climVars, type, years)

climateStacksByPeriod(tifs, climVars, type, period)
}
\arguments{
\item{tifs}{TODO}

\item{climVars}{character. CLimateNA climate variables to use.
see \url{https://climatena.ca/Help2#_var} for full list of variables.}

\item{type}{character. denotes the table in the sqlite database to access.
one of 'historical', 'future', 'historical_normals', or 'future_normals'.}

\item{years}{integer. consecutive data years to use (e.g., \code{2011:2100}).}

\item{period}{character. climate normal periods to use (e.g., \code{c("1951_1980", "1981_2010")}).}
}
\value{
list of \code{SpatRaster}s
}
\description{
Climate Stacks by Year
}
