% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepClimateLayers.R
\name{whichMSYN}
\alias{whichMSYN}
\alias{whichTypes}
\title{Determine MSYN type of a climate variable}
\usage{
whichMSYN(climVars)

whichTypes(climVars)
}
\arguments{
\item{climVars}{character. CLimateNA climate variables to use.
see \url{https://climatena.ca/Help2#_var} for full list of variables.}
}
\description{
Determine MSYN type of a climate variable
}
