% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepClimateLayers.R
\name{extractTimes}
\alias{extractTimes}
\title{Extract historical or future years (or periods) from a climate variables list}
\usage{
extractTimes(climateVarsList, type)
}
\arguments{
\item{climateVarsList}{Named list of lists, specifying the climate variables to extract and/or
calculate.
The name of each outer list element must be prefixed by either \code{future_} or \code{historical_},
and each inner list should consist of the following named elements:
\itemize{
\item \code{vars}: the raw variables used to derive the target variable;
\item \code{fun}: a quoted function used to derive the target variable,
where \code{quote(calcAsIs)} denotes target variables that ARE the raw variable:
\item \code{.dots}: additional arguments passed to \code{fun}.
}

See examples.}

\item{type}{character string specifying one of:
'historical_period', 'historical_years', 'future_period', 'future_years'.}
}
\description{
Extract historical or future years (or periods) from a climate variables list
}
