% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chelsa.R
\name{get_chelsa}
\alias{get_chelsa}
\title{Get CHELSA climate data}
\usage{
get_chelsa(type = "bioclim", layer = 1:19, period, model_string,
  scenario_string, future_years, output_dir)
}
\arguments{
\item{type}{Character. Currently only "bioclim".}

\item{layer}{Numeric. Select which bioclim layer(s) is/are downloaded.
You can choose to download any of 19 bioclimatic layers. Default is all layers (1:19).
To download bioclim layer 1, use \code{layer = 1}, or use \code{layer = 1:19} to download all layers as a rasterstack. . See details \strong{ADD LATER!}}

\item{period}{Character. Which time period to download for climate layers. One in c("past", "current", "future").}

\item{model_string}{Character. Which climatic model to download for past or future period. Only available if \code{period} is one of c("past", "future").
See \code{climatedata::check_models()} for available options.}

\item{scenario_string}{Character. Which climate scenario to download. Available options are c("rcp26", "rcp45", "rcp60", "rcp85", "pmip3"). RCP scenarios are only available if \code{period} = "future", and PMIP3 is available if \code{period} = "past".}

\item{future_years}{Character. Which time period to download for future scenario.
Available options are c("2041-2060", "2061-2080") for years 2050 and 2070.}

\item{output_dir}{Character. Path indicating the directory in which the downloaded files will be stored. Default is the current working directory.}
}
\value{
Raster* object or NULL. See \code{return_raster} argument.
}
\description{
Download data from CHELSA climatologies. This function retrieves files from the server via dowload request.
}
\examples{
output_dir <- milkunize("Projects/Crete/data-raw/Chelsa")
chelsa_bioclim <- get_chelsa(layer = 1:19, output_dir = output_dir)
}
