% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimateNA_useTiles.R
\name{getNormalsPeriods}
\alias{getNormalsPeriods}
\title{Get historical or future climate normals periods}
\usage{
getNormalsPeriods(type = NULL, tile = NULL, msy = "Y", gcm = NULL, ssp = NULL)
}
\arguments{
\item{type}{character. denotes the table in the sqlite database to access.
one of 'historical', 'future', 'historical_normals', or 'future_normals'.}

\item{tile}{integer. tile id (e.g., the output of \code{tileID()}).}

\item{msy}{character. one of 'MSY', 'M', or 'Y', corresponding to all, monthly, or yearly data.}

\item{gcm}{character. name of GCM available in ClimateNA.}

\item{ssp}{integer. one of the available SSPs in ClimateNA (e.g., \code{370} or \code{585})}
}
\value{
character.
}
\description{
Get historical or future climate normals periods
}
