% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iod.R
\name{load_iod}
\alias{load_iod}
\title{Returns Indian Ocean Dipole data. Caches it into filename unless filename is NULL.}
\usage{
load_iod(filename = "iod.rds", data_source = "jamstec", ...)
}
\description{
Returns Indian Ocean Dipole data. Caches it into filename unless filename is NULL.
}

