% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sam.R
\name{load_sam}
\alias{load_sam}
\title{Returns Southern Annular Mode data. Caches it into filename unless filename is NULL.}
\usage{
load_sam(filename = "sam.rds", data_source = "hadslp2", ...)
}
\description{
Returns Southern Annular Mode data. Caches it into filename unless filename is NULL.
}

