% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soi.R
\name{load_soi}
\alias{load_soi}
\title{Returns Southern Ocean Oscillation data. Caches it into filename unless filename is NULL.}
\usage{
load_soi(filename = "soi.rds", data_source = "bom", ...)
}
\description{
Returns Southern Ocean Oscillation data. Caches it into filename unless filename is NULL.
}
