% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLandRCS.R
\name{makeLandRCS}
\alias{makeLandRCS}
\title{Make climate normals for \code{LandR.CS}}
\usage{
makeLandRCS(
  pathToNormalRasters,
  pathToFutureRasters,
  rasterPrefix,
  outputDir,
  origTemplate,
  years = 2011:2100,
  writeCMINormal = FALSE
)
}
\arguments{
\item{pathToNormalRasters}{file path of the directory containing climate normal data,
assumed 1950-1980 and 1980-2010}

\item{pathToFutureRasters}{the file path of the directory containing the projected climate files}

\item{rasterPrefix}{name of raster file, to be followed by e.g., \file{ATA_2011-2100.grd}.}

\item{outputDir}{file path to directory where raster stacks will be written}

\item{origTemplate}{the original DEM used to create the climate files}

\item{years}{years included in the climate data}

\item{writeCMINormal}{write the CMI normal to disk}
}
\value{
\code{NULL} invisibly. Invoked for side effect of writing rasters to disk.
}
\description{
Trims and masks the climate normals, in case they aren't the same extent.
It assumes climate normal and projected data has already been generated by \code{ClimateNA}.
}
\author{
Ian Eddy, Alex Chubaty
}
