% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatedLandRCS.R
\name{makeLandRCS_projectedCMIandATA}
\alias{makeLandRCS_projectedCMIandATA}
\alias{makeLandRCS_projectedCMI}
\alias{makeLandRCS_projectedATA}
\title{Create projected CMI and ATA - the annual temperature anomaly}
\usage{
makeLandRCS_projectedCMIandATA(
  normalMAT,
  pathToFutureRasters,
  years = 2011:2100,
  studyArea = NULL
)

makeLandRCS_projectedCMI(
  normalMAT,
  pathToFutureRasters,
  years = 2011:2100,
  studyArea = NULL
)

makeLandRCS_projectedATA(
  normalMAT,
  pathToFutureRasters,
  years = 2011:2100,
  studyArea = NULL
)
}
\arguments{
\item{normalMAT}{a raster representing normal MAT}

\item{pathToFutureRasters}{directory of (annual) projected climate layers}

\item{years}{the projection years (e.g. 2011)}

\item{studyArea}{an \code{sf} polygon delineating the boundaries of the study area.
typically used to crop/mask rasters when \code{rasterToMatch} is not supplied.}
}
\value{
a list of projected rasters - CMI and ATA
}
\description{
Calculates \code{CMI} as \code{MAP - Eref}
}
\author{
Ian Eddy
}
