% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatedLandRCS.R
\name{makeLandRCS_1950_2010_normals}
\alias{makeLandRCS_1950_2010_normals}
\title{Create 1951-2010 normals as arithmetic mean of 1951-1980 and 1981-2010 normals}
\usage{
makeLandRCS_1950_2010_normals(pathToNormalRasters, rasterToMatch = NULL)
}
\arguments{
\item{pathToNormalRasters}{file path of the directory containing climate normal data,
assumed 1951-1980 and 1981-2010}

\item{rasterToMatch}{an optional template raster}
}
\value{
a \code{SpatRaster} of 1951-2010 normal CMI and MAT
}
\description{
Calculates \code{CMI} as \code{MAP - Eref}.
}
\author{
Ian Eddy
}
