% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMDC.R
\name{makeMDC}
\alias{makeMDC}
\title{Create raster of Monthly Drought Code (MDC)}
\usage{
makeMDC(
  srcdir,
  dstdir,
  tile = NULL,
  years = NULL,
  droughtMonths = 4:9,
  cl = NULL
)
}
\arguments{
\item{srcdir, dstdir}{character. vector of directory paths to source and destination raster files.}

\item{tile}{integer. tile id (e.g., the output of \code{tileID()}).}

\item{years}{integer. consecutive data years to use (e.g., \code{2011:2100}).}

\item{droughtMonths}{integer. months to calculate Monthly Drought Code (MDC). Must be \code{4:9}.}

\item{cl}{cluster object. if \code{NULL} (default), a cluster will be created using up to
\code{length(climVar)} number of CPU cores for parallel computation.
Users can pass their own \code{cl} object or specify option
\code{parallelly.availableCores.fallback} to reduce the number of cores used.
See \code{?parallelly.options}.}
}
\value{
A list of all years, with each year being the local path for the raster stack that
contains all variables.
}
\description{
Create raster of Monthly Drought Code (MDC)
}
\author{
Tati Micheletti, Ian Eddy, Alex Chubaty
}
