% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepClimateLayers.R
\name{prepClimateRasters}
\alias{prepClimateRasters}
\title{Prepare climate rasters}
\usage{
prepClimateRasters(
  climateVarsList,
  srcdir,
  dstdir,
  tile = NULL,
  years = NULL,
  type = NULL,
  cl = NULL,
  studyArea = NULL,
  studyAreaName = NULL,
  rasterToMatch = NULL,
  currentModuleName = "NoModule",
  ...
)
}
\arguments{
\item{climateVarsList}{TODO}

\item{srcdir, dstdir}{character. vector of directory paths to source and destination raster files.}

\item{tile}{integer. tile id (e.g., the output of \code{tileID()}).}

\item{years}{integer. consecutive data years to use (e.g., \code{2011:2100}).}

\item{type}{character. denotes the table in the sqlite database to access.
one of 'historic', 'future', 'historic_normals', or 'future_normals'.}

\item{cl}{cluster object. if \code{NULL} (default), a cluster will be created using up to
\code{length(climVar)} number of CPU cores for parallel computation.
Users can pass their own \code{cl} object or specify option
\code{parallelly.availableCores.fallback} to reduce the number of cores used.
See \code{?parallelly.options}.}

\item{studyArea}{an \code{sf} polygon delineating the boundaries of the study area.
typically used to crop/mask rasters when \code{rasterToMatch} is not supplied.}

\item{studyAreaName}{TODO}

\item{rasterToMatch}{TODO}

\item{currentModuleName}{TODO}

\item{...}{TODO}
}
\description{
TODO
}
