% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimateNA.R
\name{tileID}
\alias{tileID}
\title{Get the tile ID of a DEM raster}
\usage{
tileID(file)
}
\arguments{
\item{file}{character. path to a DEM raster tile file.}
}
\description{
Get the tile ID of a DEM raster
}
