% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimateNA_useTiles.R
\name{whereAmI}
\alias{whereAmI}
\title{Identify ClimateNA tiles overlapping a study area}
\usage{
whereAmI(studyArea)
}
\arguments{
\item{studyArea}{}
}
\value{
integer. vector of tile IDs.
}
\description{
Identify ClimateNA tiles overlapping a study area
}
