% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVars.R
\name{calcAsIs}
\alias{calcAsIs}
\alias{calcATA}
\alias{calcCMInormal}
\alias{calcMDC}
\title{Create raster stacks of climate variables}
\usage{
calcAsIs(stacks, layers, .dots = NULL)

calcATA(stacks, layers, .dots = NULL)

calcCMInormal(stacks, layers, .dots = NULL)

calcMDC(stacks, layers, .dots = NULL)
}
\arguments{
\item{stacks}{named list of \code{SpatRaster} objects.
names should correspond to \verb{<type>_<year>} or \verb{<type>_<period>},
where \code{type} is one of "historical" or "future"; and \code{year} or \code{period}
correspond to a climate year or climate period, respectively.}

\item{layers}{layer names (matching ClimateNA variable names) to check for in
each element of \code{stacks}.}

\item{.dots}{named list of additioonal arguments to pass to \code{fun}}
}
\value{
A \code{SpatRaster} object for a single climate variable,
with layers corresponding to year or period.
}
\description{
Produce a \code{SpatRaster} object corresponding to a single climate variable,
with layers corresponding to climate years or periods.
}
\details{
\itemize{
\item \code{calcAsIs()} returns the climate variable without modification;
\item \code{calcATA()} calculates Annual Temperature Anomaly (ATA) from \code{MAT} and \code{MAT_normal};
\item \code{calcCMInormal()} calculates mean CMI over multiple normals periods;
\item \code{calcMDC()} calculates Monthly Drought Code (MDC) from \code{Tmax} and \code{PPT} for April-September;
}
}
\author{
Alex Chubaty, Ian Eddy, Tati Micheletti
}
