% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVars.R
\name{calcMDC}
\alias{calcMDC}
\title{Create raster of Monthly Drought Code (MDC)}
\usage{
calcMDC(stacks, layers, ...)
}
\arguments{
\item{stacks}{TODO}

\item{layers}{TODO}

\item{...}{additional arguments (not used).}
}
\value{
A list of all years, with each year being the local path for the raster stack that
contains all variables.
}
\description{
Create raster of Monthly Drought Code (MDC)
}
\author{
Tati Micheletti, Ian Eddy, Alex Chubaty
}
