% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateLayersSetup.R
\name{climateLayers}
\alias{climateLayers}
\title{Helper to set up the required layers for canClimateData}
\usage{
climateLayers(
  .climVars = "CMD_sm",
  historical = TRUE,
  projected = TRUE,
  fun = quote(calcAsIs),
  historicalYears = 1991:2022,
  projectedYears = 2011:2100
)
}
\arguments{
\item{.climVars}{Character vector. This should be a vector naming the climate layers
of interest, as specified in the Climate NA web page documentation. See:
\url{https://climatena.ca/Help2#_var}}

\item{historical}{Logical. If \code{TRUE}, then include historical data in the download.}

\item{projected}{Logical. If \code{TRUE}, then include projected data in the download.}

\item{fun}{A quoted function name, e.g., \code{quote(calcAsIs)}, which is the default.
No other examples have been used and may not work.}

\item{historicalYears}{Integer vector of the historical years to request. The latest year
available as tiles is given by \code{\link[=latestHistoricalYear]{latestHistoricalYear()}}.}

\item{projectedYears}{Integer vector of the projected years to request.}
}
\value{
A named list of named lists that can be passed to the \code{sim} object named
\code{climateLayers}, which is an input requirement for \code{canClimateData} SpaDES module.
}
\description{
A constructor of the list to pass to \code{canClimateData} SpaDES module, and \code{prepClimateLayers}
function.
}
\details{
The canClimateData module requires a particular format to specify which
climater layers to download. For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{climateVariables = list(
  historical_CMDsm = list(
    vars = "historical_CMD_sm",
    fun = quote(calcAsIs),
    .dots = list(historical_years = 1991:2022)
  ),
  projected_CMDsm = list(
    vars = "future_CMD_sm",
    fun = quote(calcAsIs),
    .dots = list(future_years = 2011:2100)
  )
)
}\if{html}{\out{</div>}}

The above shows that it must be a list of lists, where the names of the list elements
must have one underscore, but the variables of the same climate layers may
need two underscores. Similarly, the default function to pass to \code{prepClimateLayers}
is \code{calcAsIs}, but this is not set by default, so the list must specify this. Finally,
the hidden list element, \code{.dots} ("dot dots") must be present and must be a list
of named years for data, and it must use the prefix \code{future} when it is about
climate projections.
}
