% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimateNA_createTiles.R
\name{ClimateNA_path}
\alias{ClimateNA_path}
\title{Construct target output paths for ClimateNA data}
\usage{
ClimateNA_path(
  dataPath,
  tile = NULL,
  type = NULL,
  msy = NULL,
  gcm = NULL,
  ssp = NULL
)
}
\arguments{
\item{dataPath}{TODO}

\item{tile}{integer. tile id (e.g., the output of \code{tileID()}).}

\item{type}{character. denotes the table in the sqlite database to access.
one of 'historic', 'future', 'historic_normals', or 'future_normals'.}

\item{msy}{character. one of 'MSY', 'M', or 'Y', corresponding to all, monthly, or yearly data.}

\item{gcm}{character. name of GCM available in ClimateNA.}

\item{ssp}{integer. one of the available SSPs in ClimateNA (e.g., \code{370} or \code{585})}
}
\value{
character string representing the path to the target output directory
}
\description{
Construct target output paths for ClimateNA data
}
