% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClimateNA_useTiles.R
\name{getClimateTable}
\alias{getClimateTable}
\title{Extract a table from the ClimateNA sqlite databose}
\usage{
getClimateTable(
  type = NULL,
  tile = NULL,
  years = NULL,
  msy = NULL,
  gcm = NULL,
  ssp = NULL
)
}
\arguments{
\item{type}{character. denotes the table in the sqlite database to access.
one of 'historic', 'future', 'historic_normals', or 'future_normals'.}

\item{tile}{integer. tile id (e.g., the output of \code{tileID()}).}

\item{years}{integer. consecutive data years to use (e.g., \code{2011:2100}).}

\item{msy}{character. one of 'MSY', 'M', or 'Y', corresponding to all, monthly, or yearly data.}

\item{gcm}{character. name of GCM available in ClimateNA.}

\item{ssp}{integer. one of the available SSPs in ClimateNA (e.g., \code{370} or \code{585})}
}
\value{
\code{data.frame}
}
\description{
Extract a table from the ClimateNA sqlite databose
}
\examples{
(getClimateTable(type = "historic", tile = 1, msy = "M", years = 2011:2020))
(getClimateTable(type = "future", tile = 1, msy = "Y", years = 2051:2060,
                 gcm = "CanESM5", ssp = 370))
(getClimateTable(type = "historic_normals", tile = 1, msy = "Y"))
(getClimateTable(type = "future_normals", tile = 1, msy = "Y",
                 gcm = "CanESM5", ssp = 370)) ## none yet available
}
