% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{assessGoogle}
\alias{assessGoogle}
\title{non-exported objects and functions from other packages}
\usage{
assessGoogle(
  url,
  archive = NULL,
  targetFile = NULL,
  destinationPath = getOption("reproducible.destinationPath", "."),
  verbose = getOption("reproducible.verbose", 1),
  team_drive = NULL
)
}
\description{
non-exported objects and functions from other packages
}
\keyword{internal}
