% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeClimateDEM.R
\name{makeClimateDEM}
\alias{makeClimateDEM}
\title{Create a DEM for use with ClimateNA}
\usage{
makeClimateDEM(
  studyArea,
  arcSecRes = c(210, 210),
  bufferArcSec,
  DEMdestinationPath,
  destinationPath,
  filename2
)
}
\arguments{
\item{studyArea}{a studyArea for cropping the gtopo30 DEM}

\item{arcSecRes}{the resolution of the output dem in arc seconds}

\item{bufferArcSec}{numeric  - arc seconds by which to buffer the studyArea to ensure seamless coverage.
Recommended to be at least the same as \code{res}}

\item{DEMdestinationPath}{DESCRIPTION NEEDED}

\item{destinationPath}{the directory to save the output raster}

\item{filename2}{the name of the output DEM without the .ASC suffix}
}
\value{
a DEM in ASCII format
}
\description{
Create a DEM for use with ClimateNA
}
\author{
Ian Eddy
}
