% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeClimateDEM.R
\name{makeClimateDEM}
\alias{makeClimateDEM}
\title{Function to create a DEM for use in climateNA}
\usage{
makeClimateDEM(
  studyArea,
  arcSecRes = c(210, 210),
  bufferArcSec,
  DEMdestinationPath,
  destinationPath,
  filename2
)
}
\arguments{
\item{studyArea}{a studyArea for cropping the gtopo30 DEM}

\item{arcSecRes}{the resolution of the output dem in arc seconds}

\item{bufferArcSec}{numeric  - arc seconds by which to buffer the studyArea to ensure seamless coverage.
Recommended to be at least the same as \code{res}}

\item{destinationPath}{the directory to save the output raster}

\item{filename2}{the name of the output DEM without the .ASC suffix}

\item{dataPath}{the directory to which gtopo30Canada is downloaded}
}
\value{
a DEM in ascii format
}
\description{
Function to create a DEM for use in climateNA
}
\author{
Ian Eddy
}
