% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeClimateDEM.R
\name{makeClimateDEM}
\alias{makeClimateDEM}
\title{Create a DEM for use with \code{ClimateNA}}
\usage{
makeClimateDEM(
  studyArea,
  arcSecRes = c(180, 180),
  bufferArcSec = 180,
  DEMdestinationPath,
  destinationPath,
  filename2
)
}
\arguments{
\item{studyArea}{an \code{sf} polygon delineating the boundaries of the study area.
typically used to crop/mask rasters when \code{rasterToMatch} is not supplied.}

\item{arcSecRes}{the resolution of the output DEM in arcseconds}

\item{bufferArcSec}{numeric, arcseconds by which to buffer the \code{studyArea} to
ensure seamless coverage.
Recommended to be at least the same as \code{res}.}

\item{DEMdestinationPath}{the directory to save the downloaded DEM raster}

\item{destinationPath}{the directory to save the output \code{.asc} raster}

\item{filename2}{the name of the output DEM without the \code{.asc} suffix}
}
\value{
a DEM in ASCII format
}
\description{
\strong{Windows only} due to limitations using \code{ClimateNA} with non-Windows
created DEMs.
}
\author{
Ian Eddy
}
