% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLandRCS.R
\name{makeLandRCS}
\alias{makeLandRCS}
\title{Function to create raster stack for of Monthly Drought Code (MDC)
this function trims and masks the climate normals, in case they aren't the same extent. 
It assumes climate normal and projected data has already been generated by climateNA
NOTE: the unit used by the output ATA is the same as the input MAT, 1/10th of a degree in climateNA}
\usage{
makeLandRCS(
  pathToNormalRasters,
  pathToFutureRasters,
  rasterPrefix,
  outputDir,
  origTemplate,
  years = 2011:2100
)
}
\arguments{
\item{pathToNormalRasters}{file path of the directory containing climate normal data, assumed 1950-1980 and 1980-2010}

\item{pathToFutureRasters}{the file path of the directory containing the projected climate files}

\item{rasterPrefix}{name of raster file, to be followed by e.g. ATA_2011-2100.grd. Put climate model here}

\item{outputDir}{file path to directory where raster stacks will be written}

\item{origTemplate}{the original DEM used to create the climate files}

\item{years}{years included in the climate data}
}
\description{
Function to create raster stack for of Monthly Drought Code (MDC)
this function trims and masks the climate normals, in case they aren't the same extent. 
It assumes climate normal and projected data has already been generated by climateNA
NOTE: the unit used by the output ATA is the same as the input MAT, 1/10th of a degree in climateNA
}
\author{
Ian Eddy
}
