% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatedLandRCS.R
\name{makeLandRCS_projectedCMIandATA}
\alias{makeLandRCS_projectedCMIandATA}
\title{Create projected CMI and ATA - the annual temperature anomaly}
\usage{
makeLandRCS_projectedCMIandATA(
  normalMAT,
  pathToFutureRasters,
  years = 2011:2100
)
}
\arguments{
\item{normalMAT}{a raster representing normal MAT}

\item{pathToFutureRasters}{directory of (annual) projected climate layers}

\item{years}{the projection years (e.g. 2011)}
}
\value{
a list of projected raster stacks - CMI and ATA
}
\description{
Calculates `CMI` as `MAP - Eref`
}
\author{
Ian Eddy
}
